+++ Robust.Benchmarks/EntityManager/HasComponentBenchmark.cs
@@ -39,7 +39,7 @@ public partial class HasComponentBenchmark
 
         for (var i = 0; i < N; i++)
         {
-            var uid = _entityManager.SpawnEntity(null, coords);
+            var uid = _entityManager.SpawnAttachedTo(null, coords);
             _entityManager.AddComponent<A>(uid);
         }
     }